+++ cp2-java-2tds-main/src/main/java/br/com/fiap/cp_api_rest/repository/PokemonRepository.java
@@ -3,5 +3,5 @@ package br.com.fiap.cp_api_rest.repository;
 import br.com.fiap.cp_api_rest.entity.Pokemon;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonRepository extends JpaRepository<Pokemon,Integer> {
+public interface PokemonRepository extends JpaRepository<Pokemon,Long> {
 }
