+++ cp2-java-2tds-main/src/main/java/br/com/fiap/cp_api_rest/repository/PokemonTrainerRepository.java
@@ -3,5 +3,5 @@ package br.com.fiap.cp_api_rest.repository;
 import br.com.fiap.cp_api_rest.entity.PokemonTrainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokemonTrainerRepository extends JpaRepository<PokemonTrainer, Integer> {
+public interface PokemonTrainerRepository extends JpaRepository<PokemonTrainer, Long> {
 }
