+++ cp2-java-2tds-main/src/main/java/br/com/fiap/cp_api_rest/repository/TrainerRepository.java
@@ -3,5 +3,5 @@ package br.com.fiap.cp_api_rest.repository;
 import br.com.fiap.cp_api_rest.entity.Trainer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainerRepository extends JpaRepository<Trainer, Integer> {
+public interface TrainerRepository extends JpaRepository<Trainer, Long> {
 }
